Modularize bactria's functionality into different plugins. Fixes #13
